unixkit_path_starts_with_any: fix correctness issue
Suppose that at some iteration the path P and its predecessor P'
are of the form

P  = X '/' Y
P' = X  c  Z

where c < '/'. In this case, the next iteration should use X as P,
rather than the longest prefix of X ending with a '/', because X is an
ancestor of P and X < P'.